+++ pywatts/core/step_information.py
@@ -1,6 +1,5 @@
 from pywatts.core.base_step import BaseStep
 
-
 class StepInformation:
     """
     This steps contains information necesary for creating a pipeline and steps by the step factory
@@ -23,3 +22,9 @@ class StepInformation:
             return StepInformation(result_step, self.pipeline)
         else:
             return self
+
+
+class SummaryInformation:
+    def __init__(self, step, pipeline):
+        self.step = step
+        self.pipeline = pipeline