+++ pywatts/summaries/__init__.py
@@ -1,0 +1,1 @@
+from pywatts.summaries.rmse_summary import RMSE